Handle initializing 64-bit ints from 32-bit constants

+++ src/revlanguage/datatypes/basic/Natural.cpp
@@ -18,7 +18,7 @@
 using namespace RevLanguage;
 
 /** Default constructor */
-Natural::Natural( void ) : Integer( 0LL )
+Natural::Natural( void ) : Integer( 0 )
 {
 
 }
